+++ src/utils/projectService/getProjectById.go
@@ -6,13 +6,11 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/zeropsio/zerops-go/errorCode"
-
-	"github.com/zeropsio/zerops-go/apiError"
-
 	"github.com/zerops-io/zcli/src/i18n"
 	"github.com/zerops-io/zcli/src/utils/sdkConfig"
+	"github.com/zeropsio/zerops-go/apiError"
 	"github.com/zeropsio/zerops-go/dto/input/path"
+	"github.com/zeropsio/zerops-go/errorCode"
 	"github.com/zeropsio/zerops-go/sdk"
 	"github.com/zeropsio/zerops-go/sdkBase"
 	"github.com/zeropsio/zerops-go/types/uuid"